Add selected all plan in trashcan

+++ Windows_Programming/Windows_Programming/Helpers/helps.cs
@@ -161,8 +161,8 @@ namespace Windows_Programming.Helpers
             var activity = new Model.Activity();
             if (documentData.TryGetValue("id", out var idValue) && idValue is long id)
                 activity.Id = (int)id;
-            if (documentData.TryGetValue("type", out var typeValue) && typeValue is int type)
-                activity.Type = type;
+            if (documentData.TryGetValue("type", out var typeValue) && typeValue is long type)
+                activity.Type = (int)type;
             if (documentData.TryGetValue("name", out var nameValue) && nameValue is string name)
                 activity.Name = name;
             if (documentData.TryGetValue("startdate", out var startDateValue) && startDateValue is string startDateString)
@@ -189,8 +189,8 @@ namespace Windows_Programming.Helpers
             var activity2 = new Transport();
             if (documentData.TryGetValue("id", out var idValue) && idValue is long id)
                 activity2.Id = (int)id;
-            if (documentData.TryGetValue("type", out var typeValue) && typeValue is int type)
-                activity2.Type = type;
+            if (documentData.TryGetValue("type", out var typeValue) && typeValue is long type)
+                activity2.Type = (int)type;
             if (documentData.TryGetValue("name", out var nameValue) && nameValue is string name)
                 activity2.Name = name;
             if (documentData.TryGetValue("startdate", out var startDateValue) && startDateValue is string startDateString)
@@ -217,8 +217,8 @@ namespace Windows_Programming.Helpers
             var activity3 = new Lodging();
             if (documentData.TryGetValue("id", out var idValue) && idValue is long id)
                 activity3.Id = (int)id;
-            if (documentData.TryGetValue("type", out var typeValue) && typeValue is int type)
-                activity3.Type = type;
+            if (documentData.TryGetValue("type", out var typeValue) && typeValue is long type)
+                activity3.Type = (int)type;
             if (documentData.TryGetValue("name", out var nameValue) && nameValue is string name)
                 activity3.Name = name;
             if (documentData.TryGetValue("startdate", out var startDateValue) && startDateValue is string startDateString)
@@ -243,8 +243,8 @@ namespace Windows_Programming.Helpers
             var activity4 = new Extend();
             if (documentData.TryGetValue("id", out var idValue) && idValue is long id)
                 activity4.Id = (int)id;
-            if (documentData.TryGetValue("type", out var typeValue) && typeValue is int type)
-                activity4.Type = type;
+            if (documentData.TryGetValue("type", out var typeValue) && typeValue is long type)
+                activity4.Type = (int)type;
             if (documentData.TryGetValue("name", out var nameValue) && nameValue is string name)
                 activity4.Name = name;
             if (documentData.TryGetValue("startdate", out var startDateValue) && startDateValue is string startDateString)
